Refine code and add in option to request a specific amount of space

+++ qc_isolates/modules/gtdbtk.nf
@@ -1,11 +1,9 @@
 process GTDBTK {
     tag "${meta.ID}"
-
-    def mem_label = params.temp_file_storage ? 'mem_16' : 'mem_120'
     
     label "cpu_8"
     label "time_12"
-    label mem_label
+    label params.temp_file_storage ? "mem_16" : "mem_120"
 
     container 'quay.io/biocontainers/gtdbtk:2.4.1--pyhdfd78af_1'
 
@@ -18,37 +16,31 @@ process GTDBTK {
     tuple val(meta), path(report_tsv), emit: results
 
     script:
-    def scratch_dir_base = params.temp_file_storage
-    def use_scratch = (scratch_dir_base != null)
-
-    def scratch_setup = use_scratch ? """
-        [[ -d "${scratch_dir_base}" ]] || { echo "Creating scratch base: ${scratch_dir_base}"; mkdir -p "${scratch_dir_base}"; }
-        SCRATCH_DIR=\$(mktemp -d -p "${scratch_dir_base}" gtdbtk_temp_XXXXXXXX)
-        echo "GTDB-Tk scratch dir: \$SCRATCH_DIR" >&2
-        cleanup() { rm -rf "\$SCRATCH_DIR"; }
-        trap cleanup EXIT
-
-    """ : """
-        echo "GTDB-Tk running WITHOUT --scratch_dir" >&2
-        SCRATCH_DIR=""
-    """
-
     report_tsv = "${meta.ID}_gtdbtk_summary.tsv"
 
     """
-    set -euo pipefail
-    export GTDBTK_DATA_PATH="${params.gtdbtk_db}"
-
-    ${scratch_setup}
+    if [[ ! -z ${params.temp_file_storage} ]]; then
+        [[ -d "${params.temp_file_storage}" ]] || { echo "Creating scratch base: ${params.temp_file_storage}"; mkdir -p "${params.temp_file_storage}"; }
+        SCRATCH_DIR=\$(mktemp -d -p "${params.temp_file_storage}" gtdbtk_temp_XXXXXXXX)
+        echo "GTDB-Tk scratch dir: \$SCRATCH_DIR" >&2
+        cleanup() { rm -rf "\$SCRATCH_DIR"; }
+        trap cleanup EXIT
+    else
+        echo "GTDB-Tk running WITHOUT --scratch_dir" >&2
+        SCRATCH_DIR=""
+    fi
 
     scratch_flag=""
     if [ -n "\$SCRATCH_DIR" ]; then
       scratch_flag="--scratch_dir \$SCRATCH_DIR"
     fi
 
+    export GTDBTK_DATA_PATH="${params.gtdbtk_db}"
+
     gtdbtk classify_wf --genome_dir fastas -x ${params.fasta_ext} --skip_ani_screen --cpus ${task.cpus} --out_dir gtdbtk_outdir \$scratch_flag
 
     cp gtdbtk_outdir/gtdbtk.bac*.summary.tsv ${report_tsv}
+
     """
 
 }
